feat: add change set pollers
Having a commit timestamp column directly in a data table that is updated
during each transaction can be a breaking change for some applications, as
a commit timestamp column that is updated using the PENDING_COMMIT_TIMESTAMP()
function will cause the owning table to no longer be readable during the
remainder of the transaction. In addition to this limitation, it is also
hard to efficiently index a commit timestamp column, which again can make
polling a data table for changes inefficient.

This change allows users to create a change set table that keeps track of
the change sets (transactions) that have been executed. Each table that is
to be watched for changes reference these change sets through an id column.
This id column can easily be indexed, as it does not contain a monotonically
increase value. The change watcher polls the change set table for changes.
When a change set is detected, the actual data table is queried for the changes
linked with the specific change set.

Fixes #52