feat: support array values for FixedShardProvider
Support array values for FixedShardProvider to allow a watcher to watch
multiple shards of a table, or even all shards. Specifying a fixed list
of shard values to use is more efficient than not specifying any values
at all for a shard column that is included in a secondary index.